requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pandas==2.3.3
numpy==2.3.3
scikit-learn==1.6.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0
